Parser can now read the entire file
Exceptions have had to be added for notes at the end of course segments.
There are also exceptions for courses that happen Every other week.
EOW's should be handled better, currently they are ignored. Notes are
also ignored. In future releases there will be a department detection
method which will allow the removal of a lot of exceptions.